refactor(ui): padroniza loading — usa crate::ui::loading::show_custom_small
- Yahoo loading: troca ui.spinner() manual por show_custom_small()
- Remove variável dark não utilizada após refactor

+++ src/app.rs
@@ -836,14 +836,11 @@ impl TemplateApp {
                                             let dates: Vec<String> = quotes
                                                 .iter()
                                                 .map(|q| {
-                                                    chrono::DateTime::from_timestamp(
-                                                        q.timestamp as i64,
-                                                        0,
-                                                    )
-                                                    .unwrap()
-                                                    .naive_utc()
-                                                    .format("%Y-%m-%d")
-                                                    .to_string()
+                                                    chrono::DateTime::from_timestamp(q.timestamp, 0)
+                                                        .unwrap()
+                                                        .naive_utc()
+                                                        .format("%Y-%m-%d")
+                                                        .to_string()
                                                 })
                                                 .collect();
                                             let df = polars::frame::DataFrame::new(vec![
